Rename kernel classes old, allow for easy new cls
