+++ src/platform/api/serviceRegistry.web.ts
@@ -4,11 +4,11 @@
 import { IInterpreterSelector } from '../interpreter/configuration/types';
 import { IInterpreterService } from '../interpreter/contracts';
 import { IServiceManager } from '../ioc/types';
-import { InterpreterSelector, InterpreterService, PythonApiProvider, PythonExtensionChecker } from './pythonApi';
+import { InterpreterSelector, InterpreterService, OldPythonApiProvider, PythonExtensionChecker } from './pythonApi';
 import { IPythonApiProvider, IPythonExtensionChecker } from './types';
 
 export function registerTypes(serviceManager: IServiceManager): void {
-    serviceManager.addSingleton<IPythonApiProvider>(IPythonApiProvider, PythonApiProvider);
+    serviceManager.addSingleton<IPythonApiProvider>(IPythonApiProvider, OldPythonApiProvider);
     serviceManager.addSingleton<IPythonExtensionChecker>(IPythonExtensionChecker, PythonExtensionChecker);
     serviceManager.addSingleton<IInterpreterService>(IInterpreterService, InterpreterService);
     serviceManager.addSingleton<IInterpreterSelector>(IInterpreterSelector, InterpreterSelector);
