+++ src/webviews/extension-side/dataviewer/dataViewerCommandRegistry.ts
@@ -28,7 +28,7 @@ import { EventName } from '../../../platform/telemetry/constants';
 import { IDataScienceErrorHandler } from '../../../kernels/errors/types';
 import { DataViewerChecker } from './dataViewerChecker';
 import { IDataViewerDependencyService, IDataViewerFactory, IJupyterVariableDataProviderFactory } from './types';
-import { PythonEnvironment } from '../../../api';
+import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
 
 @injectable()
 export class DataViewerCommandRegistry implements IExtensionSyncActivationService {
